requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-learn>=0.24.0
pandas>=1.3.0
matplotlib>=3.3.0
plotly>=5.0.0
flask>=2.0.0
pyyaml>=5.4.0
tqdm>=4.60.0
Pillow>=8.3.0
scikit-image>=0.18.0
filterpy>=1.4.5
lap>=0.4.0